Update Issue 49 CostumeAdapter and CostumeActivity Refactoring done.
This issue is now on hold, because the current version of Paintroid won't let me reproduce the problem. Waiting for the cropping feature to be implemented.

+++ catroidUiTest/src/at/tugraz/ist/catroid/uitest/content/brick/ChangeVolumeByBrickTest.java
@@ -89,7 +89,7 @@ public class ChangeVolumeByBrickTest extends ActivityInstrumentationTestCase2<Sc
 		solo.clickOnButton(0);
 
 		solo.sleep(300);
-		assertEquals("Text not updated", volumeToChange, Double.parseDouble(solo.getEditText(0).getText().toString()));
+		assertEquals("Text not updated", volumeToChange, Float.parseFloat(solo.getEditText(0).getText().toString()));
 	}
 
 	private void createProject() {
